Fixed a bug.

+++ code/PairedRead.h
@@ -33,7 +33,7 @@
  * <----------------------------------------------------------------->
  */
 class PairedRead{
-	uint64_t m_readIndex;
+	uint32_t m_readIndex;
 	uint16_t m_rank;
 	uint8_t m_library;
 public:
